add 404 page

+++ routes/shopping_cart.js
@@ -14,6 +14,7 @@ module.exports = function (app, pool) {
                 });
         } else {
             response.send([]); // this shouldn't happen
+            console.error("[ERROR] Received GET '/shopping_cart' without authenticated user.");
         }
     });
 
@@ -41,15 +42,17 @@ module.exports = function (app, pool) {
                 });
         } else {
             response.send([]); // this shouldn't happen
+            console.error("[ERROR] Received POST '/shopping_cart' without authenticated user.");
         }
     });
 
     app.delete('/shopping_cart/:id', function (request, response) {
         if (request.user) {
             let id = request.params.id;
-            let sql = "DELETE FROM shopping_cart_details WHERE id = $1";
+            let userId = request.user.id;
+            let sql = "DELETE FROM shopping_cart_details WHERE id = $1 AND shopping_cart_id = (SELECT id FROM shopping_carts WHERE user_id = $2)";
 
-            pool.query(sql, [id])
+            pool.query(sql, [id, userId])
                 .then(result => {
                     if (result.rowCount === 1) {
                         response.sendStatus(200);
@@ -65,16 +68,18 @@ module.exports = function (app, pool) {
                 });
         } else {
             response.send([]); // this shouldn't happen
+            console.error(`[ERROR] Received DELETE '/shopping_cart/${request.params.id}' without authenticated user.`);
         }
     });
 
     app.patch('/shopping_cart/:id', function (request, response) {
         if (request.user) {
-            let id = request.params.id;
             let newQuantity = request.body.quantity;
-            let sql = "UPDATE shopping_cart_details SET quantity = $1 WHERE id = $2";
+            let id = request.params.id;
+            let userId = request.user.id;
+            let sql = "UPDATE shopping_cart_details SET quantity = $1 WHERE id = $2 AND shopping_cart_id = (SELECT id FROM shopping_carts WHERE user_id = $3)";
 
-            pool.query(sql, [newQuantity, id])
+            pool.query(sql, [newQuantity, id, userId])
                 .then(result => {
                     if (result.rowCount === 1) {
                         response.sendStatus(200);
@@ -90,6 +95,26 @@ module.exports = function (app, pool) {
                 });
         } else {
             response.send([]); // this shouldn't happen
+            console.error(`[ERROR] Received PATCH '/shopping_cart/${request.params.id}' without authenticated user.`);
+        }
+    });
+
+    app.get('/checkout', function (request, response) {
+        if (request.user) {
+            let userId = request.user.id;
+            let sql = "SELECT shopping_cart_details.id, albums.id AS album_id, albums.title, quantity, albums.price FROM shopping_carts INNER JOIN shopping_cart_details ON shopping_carts.id = shopping_cart_details.shopping_cart_id INNER JOIN albums ON shopping_cart_details.album_id = albums.id WHERE shopping_carts.user_id = $1";
+
+            pool.query(sql, [userId])
+                .then(result => {
+                    response.send(result.rows);
+                })
+                .catch(e => {
+                    response.status(500).send("Database error occurred, items in cart may not be saved, please refresh or contact hectorcaesar@hotmail.com.");
+                    console.error('[ERROR] Query error:', e.message, e.stack);
+                });
+        } else {
+            response.send(); // this shouldn't happen
+            console.error("[ERROR] Received GET '/shopping_cart' without authenticated user.");
         }
     });
 };
